+++ lib/commercetools-api/src/Models/Message/ProductTailoringCreatedMessageBuilder.php
@@ -18,6 +18,7 @@ use Commercetools\Api\Models\Common\Reference;
 use Commercetools\Api\Models\Common\ReferenceBuilder;
 use Commercetools\Api\Models\Product\ProductReference;
 use Commercetools\Api\Models\Product\ProductReferenceBuilder;
+use Commercetools\Api\Models\ProductTailoring\ProductTailoringAttributeCollection;
 use Commercetools\Api\Models\ProductTailoring\ProductVariantTailoringCollection;
 use Commercetools\Api\Models\Store\StoreKeyReference;
 use Commercetools\Api\Models\Store\StoreKeyReferenceBuilder;
@@ -159,6 +160,12 @@ final class ProductTailoringCreatedMessageBuilder implements Builder
      * @var ?ProductVariantTailoringCollection
      */
     private $variants;
+
+    /**
+
+     * @var ?ProductTailoringAttributeCollection
+     */
+    private $attributes;
 
     /**
 
@@ -396,6 +403,18 @@ final class ProductTailoringCreatedMessageBuilder implements Builder
     public function getVariants()
     {
         return $this->variants;
+    }
+
+    /**
+     * <p>Attributes of the tailored Product.
+     * If available, these Attributes are selectively merged into the <code>attributes</code> of the corresponding <a href="ctp:api:type:Product">Product</a>. If the Product contains an Attribute with the same <code>name</code>, then its <code>value</code> is overwritten. Otherwise, the Attribute and its <code>value</code> are added to the Product.</p>
+     *
+
+     * @return null|ProductTailoringAttributeCollection
+     */
+    public function getAttributes()
+    {
+        return $this->attributes;
     }
 
     /**
@@ -641,6 +660,17 @@ final class ProductTailoringCreatedMessageBuilder implements Builder
     }
 
     /**
+     * @param ?ProductTailoringAttributeCollection $attributes
+     * @return $this
+     */
+    public function withAttributes(?ProductTailoringAttributeCollection $attributes)
+    {
+        $this->attributes = $attributes;
+
+        return $this;
+    }
+
+    /**
      * @param ?bool $published
      * @return $this
      */
@@ -807,6 +837,7 @@ final class ProductTailoringCreatedMessageBuilder implements Builder
             $this->metaDescription instanceof LocalizedStringBuilder ? $this->metaDescription->build() : $this->metaDescription,
             $this->metaKeywords instanceof LocalizedStringBuilder ? $this->metaKeywords->build() : $this->metaKeywords,
             $this->variants,
+            $this->attributes,
             $this->published
         );
     }