+++ lib/commercetools-api/src/Models/ProductTailoring/ProductTailoringDataBuilder.php
@@ -65,6 +65,12 @@ final class ProductTailoringDataBuilder implements Builder
     private $variants;
 
     /**
+
+     * @var ?ProductTailoringAttributeCollection
+     */
+    private $attributes;
+
+    /**
      * <p>Tailored name of the Product.</p>
      *
 
@@ -140,6 +146,18 @@ final class ProductTailoringDataBuilder implements Builder
     public function getVariants()
     {
         return $this->variants;
+    }
+
+    /**
+     * <p>Attributes of the tailored Product.
+     * If available, these Attributes are selectively merged into the <code>attributes</code> of the corresponding <a href="ctp:api:type:Product">Product</a>. If the Product contains an Attribute with the same <code>name</code>, then its <code>value</code> is overwritten. Otherwise, the Attribute and its <code>value</code> are added to the Product.</p>
+     *
+
+     * @return null|ProductTailoringAttributeCollection
+     */
+    public function getAttributes()
+    {
+        return $this->attributes;
     }
 
     /**
@@ -220,6 +238,17 @@ final class ProductTailoringDataBuilder implements Builder
     }
 
     /**
+     * @param ?ProductTailoringAttributeCollection $attributes
+     * @return $this
+     */
+    public function withAttributes(?ProductTailoringAttributeCollection $attributes)
+    {
+        $this->attributes = $attributes;
+
+        return $this;
+    }
+
+    /**
      * @deprecated use withName() instead
      * @return $this
      */
@@ -294,7 +323,8 @@ final class ProductTailoringDataBuilder implements Builder
             $this->metaDescription instanceof LocalizedStringBuilder ? $this->metaDescription->build() : $this->metaDescription,
             $this->metaKeywords instanceof LocalizedStringBuilder ? $this->metaKeywords->build() : $this->metaKeywords,
             $this->slug instanceof LocalizedStringBuilder ? $this->slug->build() : $this->slug,
-            $this->variants
+            $this->variants,
+            $this->attributes
         );
     }