+++ lib/commercetools-api/src/Models/ProductTailoring/ProductTailoringInStoreDraftBuilder.php
@@ -85,6 +85,12 @@ final class ProductTailoringInStoreDraftBuilder implements Builder
     private $variants;
 
     /**
+
+     * @var ?ProductTailoringAttributeCollection
+     */
+    private $attributes;
+
+    /**
      * <p>User-defined unique identifier of the ProductTailoring.</p>
      *
 
@@ -193,6 +199,18 @@ final class ProductTailoringInStoreDraftBuilder implements Builder
     public function getVariants()
     {
         return $this->variants;
+    }
+
+    /**
+     * <p>Attributes of the tailored Product.
+     * If provided, these Attributes are selectively merged into the <code>attributes</code> of the corresponding <a href="ctp:api:type:Product">Product</a>. If the Product contains an Attribute with the same <code>name</code>, then its <code>value</code> is overwritten. Otherwise, the Attribute and its <code>value</code> are added to the Product.</p>
+     *
+
+     * @return null|ProductTailoringAttributeCollection
+     */
+    public function getAttributes()
+    {
+        return $this->attributes;
     }
 
     /**
@@ -306,6 +324,17 @@ final class ProductTailoringInStoreDraftBuilder implements Builder
     }
 
     /**
+     * @param ?ProductTailoringAttributeCollection $attributes
+     * @return $this
+     */
+    public function withAttributes(?ProductTailoringAttributeCollection $attributes)
+    {
+        $this->attributes = $attributes;
+
+        return $this;
+    }
+
+    /**
      * @deprecated use withProduct() instead
      * @return $this
      */
@@ -394,7 +423,8 @@ final class ProductTailoringInStoreDraftBuilder implements Builder
             $this->metaKeywords instanceof LocalizedStringBuilder ? $this->metaKeywords->build() : $this->metaKeywords,
             $this->slug instanceof LocalizedStringBuilder ? $this->slug->build() : $this->slug,
             $this->publish,
-            $this->variants
+            $this->variants,
+            $this->attributes
         );
     }