+++ lib/commercetools-api/src/Models/ProductTailoring/ProductTailoringSetProductAttributeActionModel.php
@@ -6,7 +6,7 @@ declare(strict_types=1);
  * Do not change it.
  */
 
-namespace Commercetools\Api\Models\Product;
+namespace Commercetools\Api\Models\ProductTailoring;
 
 use Commercetools\Base\DateTimeImmutableCollection;
 use Commercetools\Base\JsonObject;
@@ -17,26 +17,14 @@ use stdClass;
 /**
  * @internal
  */
-final class ProductSetAttributeActionModel extends JsonObjectModel implements ProductSetAttributeAction
+final class ProductTailoringSetProductAttributeActionModel extends JsonObjectModel implements ProductTailoringSetProductAttributeAction
 {
-    public const DISCRIMINATOR_VALUE = 'setAttribute';
+    public const DISCRIMINATOR_VALUE = 'setProductAttribute';
     /**
      *
      * @var ?string
      */
     protected $action;
-
-    /**
-     *
-     * @var ?int
-     */
-    protected $variantId;
-
-    /**
-     *
-     * @var ?string
-     */
-    protected $sku;
 
     /**
      *
@@ -61,15 +49,11 @@ final class ProductSetAttributeActionModel extends JsonObjectModel implements Pr
      * @psalm-suppress MissingParamType
      */
     public function __construct(
-        ?int $variantId = null,
-        ?string $sku = null,
         ?string $name = null,
         $value = null,
         ?bool $staged = null,
         ?string $action = null
     ) {
-        $this->variantId = $variantId;
-        $this->sku = $sku;
         $this->name = $name;
         $this->value = $value;
         $this->staged = $staged;
@@ -95,46 +79,6 @@ final class ProductSetAttributeActionModel extends JsonObjectModel implements Pr
     }
 
     /**
-     * <p>The <code>id</code> of the ProductVariant to update.</p>
-     *
-     *
-     * @return null|int
-     */
-    public function getVariantId()
-    {
-        if (is_null($this->variantId)) {
-            /** @psalm-var ?int $data */
-            $data = $this->raw(self::FIELD_VARIANT_ID);
-            if (is_null($data)) {
-                return null;
-            }
-            $this->variantId = (int) $data;
-        }
-
-        return $this->variantId;
-    }
-
-    /**
-     * <p>The <code>sku</code> of the ProductVariant to update.</p>
-     *
-     *
-     * @return null|string
-     */
-    public function getSku()
-    {
-        if (is_null($this->sku)) {
-            /** @psalm-var ?string $data */
-            $data = $this->raw(self::FIELD_SKU);
-            if (is_null($data)) {
-                return null;
-            }
-            $this->sku = (string) $data;
-        }
-
-        return $this->sku;
-    }
-
-    /**
      * <p>Name of the Attribute to set.</p>
      *
      *
@@ -155,16 +99,17 @@ final class ProductSetAttributeActionModel extends JsonObjectModel implements Pr
     }
 
     /**
-     * <p>Value to set for the Attribute. If empty, any existing value will be removed.</p>
-     * <p>The <a href="ctp:api:type:AttributeType">AttributeType</a> determines the format of the Attribute <code>value</code> to be provided:</p>
+     * <p>Value to set for the Attribute. If empty, then any existing value will be removed.</p>
+     * <p><a href="ctp:api:type:AttributeType">AttributeType</a> determines the format of the Attribute <code>value</code> to be provided:</p>
      * <ul>
-     * <li>For <a href="ctp:api:type:AttributeEnumType">Enum Type</a> and <a href="ctp:api:type:AttributeLocalizedEnumType">Localized Enum Type</a>, use the <code>key</code> of the <a href="ctp:api:type:AttributePlainEnumValue">Plain Enum Value</a> or <a href="ctp:api:type:AttributeLocalizedEnumValue">Localized Enum Value</a> object or the complete object as <code>value</code>.</li>
+     * <li>For <a href="ctp:api:type:AttributeEnumType">Enum Type</a> and <a href="ctp:api:type:AttributeLocalizedEnumType">Localized Enum Type</a>,
+     * use either the <code>key</code> of the <a href="ctp:api:type:AttributePlainEnumValue">Plain Enum Value</a> or <a href="ctp:api:type:AttributeLocalizedEnumValue">Localized Enum Value</a> object or the complete object as <code>value</code>.</li>
      * <li>For <a href="ctp:api:type:AttributeLocalizableTextType">Localizable Text Type</a>, use the <a href="ctp:api:type:LocalizedString">LocalizedString</a> object as <code>value</code>.</li>
      * <li>For <a href="ctp:api:type:AttributeMoneyType">Money Type</a> Attributes, use the <a href="ctp:api:type:Money">Money</a> object as <code>value</code>.</li>
      * <li>For <a href="ctp:api:type:AttributeSetType">Set Type</a> Attributes, use the entire <code>set</code> object  as <code>value</code>.</li>
-     * <li>For <a href="ctp:api:type:AttributeNestedType">Nested Type</a> Attributes, use the list of values of all Attributes of the nested Product as <code>value</code>.</li>
      * <li>For <a href="ctp:api:type:AttributeReferenceType">Reference Type</a> Attributes, use the <a href="ctp:api:type:Reference">Reference</a> object as <code>value</code>.</li>
      * </ul>
+     * <p>Tailoring of <a href="ctp:api:type:AttributeNestedType">Nested Type</a> Attributes is <strong>not supported</strong>.</p>
      *
      *
      * @return null|mixed
@@ -184,7 +129,7 @@ final class ProductSetAttributeActionModel extends JsonObjectModel implements Pr
     }
 
     /**
-     * <p>If <code>true</code>, only the staged Attribute is set. If <code>false</code>, both the current and staged Attributes are set.</p>
+     * <p>If <code>true</code>, then only the staged Attribute is set. If <code>false</code>, then both the current and staged Attributes are set.</p>
      *
      *
      * @return null|bool
@@ -203,22 +148,6 @@ final class ProductSetAttributeActionModel extends JsonObjectModel implements Pr
         return $this->staged;
     }
 
-
-    /**
-     * @param ?int $variantId
-     */
-    public function setVariantId(?int $variantId): void
-    {
-        $this->variantId = $variantId;
-    }
-
-    /**
-     * @param ?string $sku
-     */
-    public function setSku(?string $sku): void
-    {
-        $this->sku = $sku;
-    }
 
     /**
      * @param ?string $name