+++ lib/commercetools-import/src/Models/Products/ProductImportBuilder.php
@@ -24,6 +24,7 @@ use Commercetools\Import\Models\Common\StateKeyReference;
 use Commercetools\Import\Models\Common\StateKeyReferenceBuilder;
 use Commercetools\Import\Models\Common\TaxCategoryKeyReference;
 use Commercetools\Import\Models\Common\TaxCategoryKeyReferenceBuilder;
+use Commercetools\Import\Models\Productvariants\AttributeCollection;
 use stdClass;
 
 /**
@@ -66,6 +67,12 @@ final class ProductImportBuilder implements Builder
      * @var ?CategoryKeyReferenceCollection
      */
     private $categories;
+
+    /**
+
+     * @var ?AttributeCollection
+     */
+    private $attributes;
 
     /**
 
@@ -185,6 +192,15 @@ final class ProductImportBuilder implements Builder
     public function getCategories()
     {
         return $this->categories;
+    }
+
+    /**
+
+     * @return null|AttributeCollection
+     */
+    public function getAttributes()
+    {
+        return $this->attributes;
     }
 
     /**
@@ -370,6 +386,17 @@ final class ProductImportBuilder implements Builder
     public function withCategories(?CategoryKeyReferenceCollection $categories)
     {
         $this->categories = $categories;
+
+        return $this;
+    }
+
+    /**
+     * @param ?AttributeCollection $attributes
+     * @return $this
+     */
+    public function withAttributes(?AttributeCollection $attributes)
+    {
+        $this->attributes = $attributes;
 
         return $this;
     }
@@ -581,6 +608,7 @@ final class ProductImportBuilder implements Builder
             $this->slug instanceof LocalizedStringBuilder ? $this->slug->build() : $this->slug,
             $this->description instanceof LocalizedStringBuilder ? $this->description->build() : $this->description,
             $this->categories,
+            $this->attributes,
             $this->metaTitle instanceof LocalizedStringBuilder ? $this->metaTitle->build() : $this->metaTitle,
             $this->metaDescription instanceof LocalizedStringBuilder ? $this->metaDescription->build() : $this->metaDescription,
             $this->metaKeywords instanceof LocalizedStringBuilder ? $this->metaKeywords->build() : $this->metaKeywords,